+++ sample/src/main/java/com/ornoma/httpcache/sample/VideoFragment.java
@@ -8,9 +8,8 @@ import android.widget.ImageView;
 import android.widget.SeekBar;
 import android.widget.VideoView;
 
-import com.danikula.videocache.CacheListener;
-import com.danikula.videocache.HttpProxyCacheServer;
-import com.danikula.videocache.sample.R;
+import com.ornoma.httpcache.CacheListener;
+import com.ornoma.httpcache.HttpProxyCacheServer;
 
 import org.androidannotations.annotations.AfterViews;
 import org.androidannotations.annotations.EFragment;
@@ -84,10 +83,10 @@ public class VideoFragment extends Fragment implements CacheListener {
     }
 
     @Override
-    public void onCacheAvailable(File file, String url, int percentsAvailable) {
-        progressBar.setSecondaryProgress(percentsAvailable);
+    public void onCacheAvailable(File file, String url, double percentsAvailable) {
+        progressBar.setSecondaryProgress((int)percentsAvailable);
         setCachedState(percentsAvailable == 100);
-        Log.d(LOG_TAG, String.format("onCacheAvailable. percents: %d, file: %s, url: %s", percentsAvailable, file, url));
+        Log.d(LOG_TAG, String.format("onCacheAvailable. percents: %d, file: %s, url: %s", (int)percentsAvailable, file, url));
     }
 
     private void updateVideoProgress() {
